requirements.txt
aiomysql==0.1.1
aioredis==2.0.1
aiosqlite==0.17.0
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cos-python-sdk-v5==1.9.15
crcmod==1.7
cryptography==3.4.8
Deprecated==1.2.13
dicttoxml==1.7.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.75.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
loguru==0.6.0
MarkupSafe==2.1.1
optionaldict==0.1.2
orjson==3.7.7
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
prompt-toolkit==3.0.30
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.9.4
pydantic==1.9.1
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==3.0.9
pypika-tortoise==0.1.5
python-alipay-sdk==3.0.4
python-dateutil==2.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
qrcode==7.3.1
redis==4.3.4
requests==2.27.1
requests-pkcs12==1.12
sentry-sdk==1.6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.17.1
tencentcloud-sdk-python==3.0.618
tortoise-orm==0.19.0
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvicorn==0.17.6
uvloop==0.16.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
wechatpy==2.0.0a11
wrapt==1.14.1
xmltodict==0.13.0
arrow==1.2.2
alembic==1.8.1